Using non-volatile data directories so installs survive reboots.
I noticed that following `vagrant reload` that the hdfs, namenode, and
datanode services would not start. This is due to the data dir being
created within /var/run, which is mounted on a volatile filesystem. As a
result, not only will data be lost during a `vagrant reload`, but the
directory structure created during `hdsf namenode -format` is lost. This
prevents the namenode service from starting, which in turn prevents the
datanode service from starting.

jmaron found this same issue
https://github.com/hortonworks/structor/issues/5, and the change here
fixes it, I believe, by moving the data directories to /var/lib, which
is non-volatile.